Added simulator support for foxglove realtime sim
