+++ simulation/asv_simulator/config/asv_sim.yaml
@@ -1,0 +1,4 @@
+/**:
+  ros__parameters:
+    asv_sim:
+
