+++ simulation/asv_simulator/launch/asv_simulator.launch.py
@@ -1,3 +1,6 @@
+import os
+from ament_index_python.packages import get_package_share_directory
+
 from launch import LaunchDescription
 from launch.actions import DeclareLaunchArgument, LogInfo
 from launch.substitutions import LaunchConfiguration
@@ -6,7 +9,7 @@ from launch_ros.actions import Node
 def generate_launch_description():
     declare_plotting_method_arg = DeclareLaunchArgument(
         'plotting_method',
-        default_value='matplotlib',
+        default_value='foxglove',
         description='Plotting method to use: matplotlib or foxglove')
 
     asv_simulator_node = Node(
@@ -14,7 +17,9 @@ def generate_launch_description():
             executable='asv_simulator_node.py',
             name='asv_simulator_node',
             output='screen',
-            parameters=[{'plotting_method': LaunchConfiguration('plotting_method')}]
+            parameters=[
+                os.path.join(get_package_share_directory('asv_setup'), 'config', 'robots', 'freya.yaml'),
+                {'plotting_method': LaunchConfiguration('plotting_method')}]
         )
 
     return LaunchDescription([
